+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Debounce the documents and split-expenses search inputs.
+
 ## [0.57.4] - 2026-06-02
 
 ### Changed